separate realtime indexing store (#1229)
* separate realtime indexing store

* cleanup
